testutils: run roachvet on test files
The previously constructed extended regular expression would end up
looking like:

    (pkg/*_test.go)|(pkg/workload)|...|(pkg/testutil):.*Use of go keyword not allowed

As a result, roachvet was being skipped on all test files and all but
the last package in this list.

This fixes the filter and addresses the violations that had snuck into
the code:

- *: fix unnecessary conversion lint violations
- *: fix composite literal uses unkeyed fields violations
- *: fix 'format argument is not a constant expression' violations
- *: fix 'message argument is not a constant expression' violations
- *: fix 'invalid direct cast on error object' violations
- *: fix 'use errors.Is instead of a direct comparison' violations
- changefeedccl: fix returnerrcheck violations
- *: fix shadowing violations
- *: fix missing calls to context.Cancel
- kv/kvserver: set timer.Read = true
- *: fix various format string mistakes
- *: fix grpcconnclose violations
- kvtenantccl: fix unreachable code violation

Release note: None

+++ pkg/kv/kvserver/gc_queue_test.go
@@ -1169,7 +1169,7 @@ func TestGCQueueChunkRequests(t *testing.T) {
 	if err != nil {
 		t.Fatalf("could not find span config for range %s", err)
 	}
-	tc.manualClock.Increment(int64(conf.TTL().Nanoseconds()) + 1)
+	tc.manualClock.Increment(conf.TTL().Nanoseconds() + 1)
 	gcQ := newGCQueue(tc.store)
 	processed, err := gcQ.process(ctx, tc.repl, confReader)
 	if err != nil {
